Refactor pagination handling by moving functionality to tables.py and updating related templates for improved structure and usability
