--- common/pagination.py
@@ -1,22 +1,0 @@
-"""
-Pagination Related Common Functions
-"""
-
-from django.http import HttpRequest
-
-
-def pagination_handle(request: HttpRequest, default_size=10, default_page_number=1):
-    """
-    Handle Pagination Size and page Number Parameter
-    """
-    size = request.GET.get("size", str(default_size))
-    if size.isnumeric():
-        size = int(size)
-    else:
-        size = default_size
-    page_number = request.GET.get("page", str(default_page_number))
-    if page_number.isnumeric():
-        page_number = int(page_number)
-    else:
-        page_number = default_page_number
-    return size, page_number
