+++ common/templates/table.html
@@ -17,4 +17,21 @@
             {% endfor %}
         </tbody>
       </table>
+      <nav class="pagination" role="navigation" aria-label="pagination">
+        {% if table.has_previous %}
+          <a class="pagination-previous" href="?page={{ table.previous_page_number }}&size={{ table.page_size }}">Previous</a>
+        {% endif %}
+      
+        <ul class="pagination-list">
+          {% for page_num in paginator.page_range %}
+            <li style="padding-right: 10px;">
+              <a href="?page={{ page_num }}&size={{ table.page_size }}"  {% if page_num == table.page_number %}class="has-text-success"{% endif %}>{{ page_num }}</a>
+            </li>
+          {% endfor %}
+        </ul>
+      
+        {% if table.has_next %}
+          <a class="pagination-next" href="?page={{ table.next_page_number }}&size={{ table.page_size }}">Next</a>
+        {% endif %}
+      </nav>
 </div>
