+++ country/templates/countries.html
@@ -32,22 +32,6 @@
     {% with table=countries %}
       {% include 'table.html' %}
     {% endwith %}
-  <nav class="pagination" role="navigation" aria-label="pagination">
-    {% if customers.has_previous %}
-      <a class="pagination-previous" href="?page={{ customers.previous_page_number }}">Previous</a>
-    {% endif %}
   
-    <ul class="pagination-list">
-      {% for page_num in paginator.page_range %}
-        <li style="padding-right: 10px;">
-          <a href="?page={{ page_num }}"  {% if page_num == customers.number %}class="has-text-success"{% endif %}>{{ page_num }}</a>
-        </li>
-      {% endfor %}
-    </ul>
-  
-    {% if customers.has_next %}
-      <a class="pagination-next" href="?page={{ customers.next_page_number }}">Next</a>
-    {% endif %}
-  </nav>
 </div>
 {% endblock %}
